Index terrain patch cache by quad tree position instead of hashing
- Replace the cache's HashMap<PatchKey, Entry> with a flat residency table
  indexed by PatchIndex, plus a dense entry array. The quad tree layout is
  already a perfect hash over every patch, so no hashing is needed.
- Sibling residency now reads one cache line, and selection carries the atlas
  slot out with it, removing a second lookup and a runtime panic.
- Settle on one noun: PatchKey/NodeIndex/TreeNode/PatchData become
  PatchCoord/PatchIndex/TreePatch/PatchPayload.
- Move the layout into patch.rs as PatchLayout, breaking the cache <-> quad
  tree import cycle. Offsets are a pyramid, total count in the last slot.
- Split height_range out of the payload so only texels flow through the cache.
- PatchSelection.needed lists every patch the frame depends on, instead of
  deriving it by chaining two vectors.
- Guard the patch <-> entry mapping with asserts on both ends; add cache tests
  covering swap_remove relocation and atlas slot reclamation.
- Stop panicking when a selection is wider than the atlas.
- Add LOD colors 8-10, which rendered black after PATCH_LOD_COUNT went to 11.
- Name the gradient atlas contents honestly: it stores the normal's horizontal
  part, so the shader needs no sign flip.

+++ crates/app/src/terrain/texture_atlas.rs
@@ -8,7 +8,7 @@ use crate::FRAME_COUNT;
 use crate::d3d12_utils::{D3D12BufferExt, D3D12TextureExt, InterfaceExt};
 
 #[repr(transparent)]
-#[derive(Clone, Copy, PartialEq, Eq)]
+#[derive(Clone, Copy, PartialEq, Eq, Debug)]
 pub struct AtlasSlot(UVec2);
 
 impl AtlasSlot {
@@ -122,8 +122,8 @@ impl<T> TextureAtlas<T> {
                     Type: D3D12_TEXTURE_COPY_TYPE_SUBRESOURCE_INDEX,
                     Anonymous: D3D12_TEXTURE_COPY_LOCATION_0 { SubresourceIndex: 0 },
                 },
-                slot.coords().x * ATLAS_PATCH_SIZE_IN_PIXELS as u32,
-                slot.coords().y * ATLAS_PATCH_SIZE_IN_PIXELS as u32,
+                slot.coords().x * ATLAS_PATCH_SIZE_IN_PIXELS,
+                slot.coords().y * ATLAS_PATCH_SIZE_IN_PIXELS,
                 0,
                 &D3D12_TEXTURE_COPY_LOCATION {
                     pResource: std::mem::transmute_copy(&self.upload),
@@ -133,8 +133,8 @@ impl<T> TextureAtlas<T> {
                             Offset: dst_patch_base as u64,
                             Footprint: D3D12_SUBRESOURCE_FOOTPRINT {
                                 Format: self.format,
-                                Width: ATLAS_PATCH_SIZE_IN_PIXELS as u32,
-                                Height: ATLAS_PATCH_SIZE_IN_PIXELS as u32,
+                                Width: ATLAS_PATCH_SIZE_IN_PIXELS,
+                                Height: ATLAS_PATCH_SIZE_IN_PIXELS,
                                 Depth: 1,
                                 RowPitch: row_pitch as u32,
                             },
